refactor(ui): move WorksPageContent table into separate WorksTable component
