+++ app/(logged_in)/creativity/WorksPageContent.tsx
@@ -1,25 +1,15 @@
 'use client';
 
-import {
-  Accordion,
-  AccordionDetails,
-  AccordionSummary,
-  Box,
-  Button,
-  IconButton,
-  MenuItem,
-  Typography
-} from '@mui/material';
-import { ChevronDown, ChevronRight, MoreVertical } from 'lucide-react';
+import { Box, Button, MenuItem } from '@mui/material';
+import { ChevronDown } from 'lucide-react';
 import Link from 'next/link';
 import { useRef, useState } from 'react';
 
-import { StatusWithDate } from './components/StatusWithDate';
 import { useWorksFiltering } from './useWorksFiltering';
 import { type OpusGroup, type UngroupedGroup, WORKS_MOCK_DATA, type WorkStatus } from './works.mock';
-import { getContextMenuDropdownItem, getGroupedWorkRowSx, styles, TABLE_DIVIDER_COLOR } from './WorksPageContent.styles';
+import { styles } from './WorksPageContent.styles';
+import { WorksTable } from './WorksTable';
 import {
-  WORKS_BASE_PATH,
   WORKS_CREATE_OPTIONS,
   WORKS_EMPTY_STATE_DESCRIPTION,
   WORKS_EMPTY_STATE_NO_RESULTS_DESCRIPTION,
@@ -32,7 +22,6 @@ import {
   type WorksTabValue
 } from '~/constants/creativity';
 import type { FilesSortValue } from '~/constants/sort';
-import PencilIcon from '~/public/icons/pencil.svg';
 import DropdownMenu from '~/shared/components/dropdown-menu/DropdownMenu';
 import { EmptyState } from '~/shared/components/empty-state';
 import { FilteringToolbar, SortSelect } from '~/shared/components/filtering-toolbar';
@@ -78,16 +67,6 @@ function isWorksStatusValue(status: WorkStatus): status is WorksStatusValue {
   return WORKS_STATUSES.includes(status as WorksStatusValue);
 }
 
-type IndividualWorkData = Readonly<{
-  id: string;
-  title: string;
-  year: string;
-  genre: string;
-  status: WorkStatus;
-  updatedAt: string;
-  language: 'uk' | 'en' | 'bilingual';
-}>;
-
 function toGroupRowData(group: OpusGroup | UngroupedGroup): GroupRowData {
   const [rawStartDate, rawEndDate] = group.yearRange.split('-').map((part) => part.trim());
   const startDate = rawStartDate || group.updatedAt.slice(0, 4);
@@ -106,18 +85,6 @@ function toGroupRowData(group: OpusGroup | UngroupedGroup): GroupRowData {
     publishedAt: group.publishedAt,
     works: group.works
   };
-}
-
-function formatGroupYears(startDate: string, endDate?: string): string {
-  if (!endDate || endDate === startDate) {
-    return startDate;
-  }
-
-  return `${startDate} - ${endDate}`;
-}
-
-function getGroupEditHref(groupId: string): string {
-  return `${WORKS_BASE_PATH}/group/${groupId}/edit`;
 }
 
 function sortGroups<T extends { title: string; updatedAt: string }>(
@@ -140,28 +107,6 @@ function sortGroups<T extends { title: string; updatedAt: string }>(
     return right.title.localeCompare(left.title, 'uk', { sensitivity: 'base' });
   });
 }
-
-// ── Opus context menu ─────────────────────────────────────────────────────────
-const GROUP_MENU_ITEMS = [
-  { id: 'edit', label: 'Редагувати' },
-  { id: 'publish', label: 'Опублікувати' },
-  { id: 'unpublish', label: 'Зняти з публікації' },
-  { id: 'ungroup', label: 'Розгрупувати' },
-  { id: 'seo', label: 'SEO налаштування' },
-  { id: 'share', label: 'Поширити' },
-  { id: 'delete', label: 'Видалити', danger: true }
-] as const;
-
-// ── Ungrouped work context menu ───────────────────────────────────────────────
-const WORK_MENU_ITEMS = [
-  { id: 'upload_audio', label: 'Завантажити аудіо' },
-  { id: 'upload_pdf', label: 'Завантажити PDF' },
-  { id: 'seo', label: 'SEO налаштування' },
-  { id: 'share', label: 'Поширити' },
-  { id: 'delete', label: 'Видалити', danger: true }
-] as const;
-
-type ContextMenuItem = { id: string; label: string; danger?: boolean };
 
 function useDropdownState() {
   const [anchorEl, setAnchorEl] = useState<HTMLElement | null>(null);
@@ -192,185 +137,6 @@ function DropdownItemsList<T extends { id: string }>({
   renderItem: (item: T) => React.ReactNode;
 }>) {
   return <Box sx={styles.menuList}>{items.map((item) => renderItem(item))}</Box>;
-}
-
-function ContextMenu({
-  items,
-  triggerLabel
-}: Readonly<{
-  items: readonly ContextMenuItem[];
-  triggerLabel: string;
-}>) {
-  const { anchorEl, triggerRef, handleOpen, handleClose } = useDropdownState();
-
-  return (
-    <>
-      <Box sx={styles.contextMenuWrapper}>
-        <IconButton
-          component="span"
-          ref={triggerRef}
-          onClick={handleOpen}
-          aria-label={triggerLabel}
-          aria-haspopup="menu"
-          aria-expanded={Boolean(anchorEl)}
-          sx={styles.contentMenuButton}
-        >
-          <MoreVertical size={22} color="#190D03" />
-        </IconButton>
-      </Box>
-
-      <DropdownMenu
-        disableScrollLock
-        anchorEl={anchorEl}
-        open={Boolean(anchorEl)}
-        onClose={handleClose}
-        sx={styles.contextMenuDropdown}
-        anchorOrigin={{ vertical: 'bottom', horizontal: 'right' }}
-        transformOrigin={{ vertical: 'top', horizontal: 'right' }}
-        menuList={
-          <DropdownItemsList
-            items={items}
-            renderItem={(item) => (
-              <MenuItem key={item.id} onClick={handleClose} sx={getContextMenuDropdownItem(item.danger)}>
-                {item.label}
-              </MenuItem>
-            )}
-          />
-        }
-      />
-    </>
-  );
-}
-
-function EditAction({ href, label }: Readonly<{ href: string; label: string }>) {
-  const handleClick = (event: React.MouseEvent<HTMLAnchorElement>): void => {
-    event.stopPropagation();
-  };
-
-  return (
-    <Box sx={styles.editActionWrapper}>
-      <IconButton component={Link} href={href} onClick={handleClick} aria-label={label} sx={styles.editActionButton}>
-        <PencilIcon />
-      </IconButton>
-    </Box>
-  );
-}
-
-function WorkRowCells({ title, year }: Readonly<{ title: string; year: string }>) {
-  return (
-    <>
-      <Typography sx={styles.workRowTitle}>{title}</Typography>
-      <Box sx={styles.genreSpacer} />
-      <Typography sx={styles.yearText}>{year}</Typography>
-      <Box sx={styles.statusSpacer} />
-      <Box sx={styles.rowActionsCell}>
-        <ContextMenu items={WORK_MENU_ITEMS} triggerLabel={`Дії твору ${title}`} />
-      </Box>
-    </>
-  );
-}
-
-function WorksTableHeader() {
-  return (
-    <Box sx={styles.tableHeader}>
-      <Box sx={styles.markerColumn} />
-      <Typography sx={styles.tableHeaderFirstText}>Опуси</Typography>
-      <Typography sx={styles.tableHeaderText}>Назва</Typography>
-      <Typography sx={styles.tableHeaderText}>Жанр</Typography>
-      <Typography sx={styles.tableHeaderText}>Роки</Typography>
-      <Typography sx={styles.tableHeaderText}>Статус</Typography>
-      <Box sx={styles.actionsSpacer} />
-    </Box>
-  );
-}
-
-// ── Opus accordion row ────────────────────────────────────────────────────────
-function GroupedRow({
-  group,
-  defaultExpanded
-}: Readonly<{
-  group: GroupRowData;
-  defaultExpanded?: boolean;
-}>) {
-  return (
-    <Accordion defaultExpanded={defaultExpanded} disableGutters elevation={0} sx={styles.accordion}>
-      <AccordionSummary expandIcon={<ChevronRight size={18} />} sx={styles.accordionSummary}>
-        <Box sx={styles.accordionSummaryContent}>
-          <Box sx={styles.accordionSummaryContentSpacer} />
-
-          <Typography sx={styles.opusNumberTypography}>{group.numberLabel}</Typography>
-
-          <Typography sx={styles.groupTitleText}>{group.title}</Typography>
-
-          <Typography sx={styles.groupGenreText}>{group.genre}</Typography>
-
-          <Typography sx={styles.groupYearsText}>{formatGroupYears(group.startDate, group.endDate)}</Typography>
-
-          <Box sx={styles.statusColumnWrapper}>
-            <StatusWithDate
-              status={group.status}
-              createdAt={group.createdAt}
-              updatedAt={group.updatedAt}
-              publishedAt={group.publishedAt}
-              dividerColor={TABLE_DIVIDER_COLOR}
-            />
-          </Box>
-
-          <Box sx={styles.rowActionsCell}>
-            <EditAction href={getGroupEditHref(group.id)} label={`Редагувати групу ${group.title}`} />
-
-            <ContextMenu items={GROUP_MENU_ITEMS} triggerLabel={`Дії групи ${group.title}`} />
-          </Box>
-        </Box>
-      </AccordionSummary>
-
-      {group.works.length > 0 && (
-        <AccordionDetails sx={styles.accordionDetails}>
-          {group.works.map((work, index) => (
-            <Box
-              key={work.id}
-              sx={getGroupedWorkRowSx(index === group.works.length - 1)}
-            >
-              <Box sx={styles.markerColumn} />
-              <WorkRowCells title={work.title} year={work.year} />
-            </Box>
-          ))}
-        </AccordionDetails>
-      )}
-    </Accordion>
-  );
-}
-
-// ── Individual work (ungrouped work without container) ──────────────────────
-function IndividualWorkRow({ work }: Readonly<{ work: IndividualWorkData }>) {
-  return (
-    <Box sx={styles.individualWorkRow}>
-      {/* Left marker column (empty for individual works) */}
-      <Box sx={styles.markerColumn} />
-
-      <Typography sx={styles.workRowTitle}>{work.title}</Typography>
-
-      <Box sx={styles.genreSpacer} />
-
-      <Typography sx={styles.individualWorkYearText}>{work.year}</Typography>
-
-      <Box sx={styles.statusSpacer} />
-
-      <Box sx={styles.rowActionsCell}>
-        <ContextMenu items={WORK_MENU_ITEMS} triggerLabel={`Дії твору ${work.title}`} />
-      </Box>
-    </Box>
-  );
-}
-
-function GroupRowsList({ groups }: Readonly<{ groups: readonly (OpusGroup | UngroupedGroup)[] }>) {
-  return (
-    <>
-      {groups.map((group) => (
-        <GroupedRow key={group.id} group={toGroupRowData(group)} defaultExpanded />
-      ))}
-    </>
-  );
 }
 
 // ── Create button (same pattern as PublicationsCreateAction) ──────────────────
@@ -460,7 +226,6 @@ export function WorksPageContent({ activeTab }: WorksPageContentProps) {
 
   const visibleUngroupedGroups = getVisibleGroups(WORKS_MOCK_DATA.ungroupedGroups, (group) => group.boNumber);
 
-  // Extract all individual ungrouped works (when 'works' tab is active)
   const allUngroupedWorks = WORKS_MOCK_DATA.ungroupedGroups.flatMap((group) =>
     group.works.map((work) => ({
       id: work.id,
@@ -487,8 +252,11 @@ export function WorksPageContent({ activeTab }: WorksPageContentProps) {
     const showUngrouped = activeTab === 'all' || activeTab === 'ungrouped';
     const showIndividualWorks = activeTab === 'all' || activeTab === 'works';
 
-    const hasOpus = showOpus && visibleOpusGroups.length > 0;
-    const hasUngrouped = showUngrouped && visibleUngroupedGroups.length > 0;
+    const mappedOpusGroups = visibleOpusGroups.map(toGroupRowData);
+    const mappedUngroupedGroups = visibleUngroupedGroups.map(toGroupRowData);
+
+    const hasOpus = showOpus && mappedOpusGroups.length > 0;
+    const hasUngrouped = showUngrouped && mappedUngroupedGroups.length > 0;
     const hasIndividualWorks = showIndividualWorks && visibleUngroupedWorks.length > 0;
 
     if (!hasOpus && !hasUngrouped && !hasIndividualWorks) {
@@ -501,15 +269,14 @@ export function WorksPageContent({ activeTab }: WorksPageContentProps) {
     }
 
     return (
-      <Box sx={styles.worksListContainer}>
-        <WorksTableHeader />
-
-        {showOpus && <GroupRowsList groups={visibleOpusGroups} />}
-
-        {showUngrouped && <GroupRowsList groups={visibleUngroupedGroups} />}
-
-        {showIndividualWorks && visibleUngroupedWorks.map((work) => <IndividualWorkRow key={work.id} work={work} />)}
-      </Box>
+      <WorksTable
+        visibleOpusGroups={mappedOpusGroups}
+        visibleUngroupedGroups={mappedUngroupedGroups}
+        visibleUngroupedWorks={visibleUngroupedWorks}
+        showOpus={showOpus}
+        showUngrouped={showUngrouped}
+        showIndividualWorks={showIndividualWorks}
+      />
     );
   })();
